backend/requirements.txt
fastapi[standard]
openai
python-dotenv
pytest
pytest-asyncio
pytest-mock